requirements.txt
alphashape==1.1.0
appnope==0.1.0
Arpeggio==1.9.2
astroid==2.5.3
attrs==20.2.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
click-log==0.3.2
cobra==0.19.0
confluent-kafka==1.5.0
cycler==0.10.0
data-science-types==0.2.18
decorator==4.4.2
depinfo==1.5.4
ete3==3.1.2
fonttools==4.32.0
future==0.18.2
importlib-metadata==2.0.0
importlib-resources==3.0.0
iniconfig==1.0.1
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.4
jedi==0.17.2
joblib==1.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.5.1
mccabe==0.6.1
mpmath==1.1.0
mypy==0.782
mypy-extensions==0.4.3
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.2.52
optlang==1.4.6
packaging==20.4
pandas==1.1.2
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
Pint==0.19.1
pip-licenses==3.3.1
pipdeptree==1.0.0
pluggy==0.13.1
prompt-toolkit==3.0.7
PTable==0.9.2
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
Pygments==2.8.1
pylint==2.7.4
pymongo==4.1.0
pymunk==5.7.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pytest==6.1.1
python-dateutil==2.8.1
python-libsbml-experimental==5.18.1
pytz==2020.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.8.0
Shapely==1.7.1
six==1.15.0
stochastic-arrow==0.4.3
swiglpk==4.65.1
sympy==1.6.2
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.56.2
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
vivarium-core==1.2.3
vivarium-scripts==0.0.5
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
zipp==3.3.0
